+++ ndindex/tests/test_chunking.py
@@ -106,6 +106,7 @@ def test_indices(chunk_size, shape):
 def test_as_subchunks_error():
     raises(ValueError, lambda: next(ChunkSize((1, 2)).as_subchunks(..., (1, 2, 3))))
 
+@example((1,), (0,), ...)
 @example((2, 2), (5, 5), (0, 3))
 @example((2, 2), (5, 5), (slice(0, 5, 2), slice(0, 5, 3)))
 @example((2, 2), (5, 5), ([0, 0],))
@@ -127,9 +128,9 @@ def test_as_subchunks(chunk_size, shape, idx):
         subarrays = []
         fast = chunk_size.as_subchunks(idx, shape)
         slow = chunk_size.as_subchunks(idx, shape, _force_slow=True)
-        for (c, index), (cslow, indexslow) in zip_longest(fast, slow):
+        for c, cslow in zip_longest(fast, slow):
             assert c == cslow
-            assert index == indexslow
+            index = idx.as_subindex(c)
             chunk = a[c.raw]
             subchunk = chunk[index.raw]
             # Not empty